Use userId istead of email
When I was doing the deserialization of the user, it didn't allow me to use the id, it was because I was doing a req.body and obviously the id didn't appear when you wanted to register or login. Now from the database to the pages, they use the user id instead of the email.